Move featureconfig and capacityplanner packages to base dir
