+++ pkg/base/capacityplanner/common.go
