+++ pkg/base/capacityplanner/helpers.go
